Modulariza record.js en módulos ES dentro de src/
Divide el archivo monolítico de 404 líneas en módulos por responsabilidad
(state, recorder, player, transcription, pagination, nav, main) sin alterar
el comportamiento existente, y elimina console.log y código muerto de debug.
Prerequisito para la implementación del afinador (NV-002).
