+++ index.html
@@ -56,6 +56,6 @@
       lucide.createIcons();
     </script>
     <script src="node_modules/toastify-js/src/toastify.js"></script>
-    <script src="record.js"></script>
+    <script type="module" src="src/main.js"></script>
   </body>
 </html>
