+++ src/state.js
@@ -1,0 +1,10 @@
+export const AUDIOS_PER_PAGE = 5;
+
+export const recordedAudios = [];
+
+export const state = {
+  currentPage: 1,
+  draggingElement: null,
+  draggingAudioIndex: null,
+  transcriptsViewOpen: false,
+};
